Isolate Qwen3-VL two-turn NPU test
